refactoring blog app
