+++ apps/blog/apps.py
@@ -1,15 +1,16 @@
 from django.apps import AppConfig
 
 
-class BlogConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'blog'
+# class BlogConfig(AppConfig):
+#     default_auto_field = 'django.db.models.BigAutoField'
+#     name = 'blog'
 
 
    
 class BlogConfig(AppConfig):
-    name = 'blog'
+    name = 'apps.blog'
+    default_auto_field = 'django.db.models.BigAutoField'
 
     def ready(self):
-        import blog.signals
+        import apps.blog.signals
