--- apps/blog/models.py
@@ -1,41 +1,0 @@
-from django.db import models
-
-class Tag(models.Model):
-    name = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.name
-
-class Category(models.Model):
-    name = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.name
-
-class Author(models.Model):
-    name = models.CharField(max_length=100)
-    bio = models.TextField()
-
-    def __str__(self):
-        return self.name
-
-class Blog(models.Model):
-    title = models.CharField(max_length=200)
-    content = models.TextField()
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    author = models.ForeignKey(Author, on_delete=models.CASCADE)
-    categories = models.ManyToManyField(Category)
-    tags = models.ManyToManyField(Tag)
-
-    def __str__(self):
-        return self.title
-
-class Image(models.Model):
-    image_file = models.ImageField(upload_to='images/')
-    uploaded_at = models.DateTimeField(auto_now_add=True)
-    blog_post = models.ForeignKey(Blog, related_name='images', on_delete=models.CASCADE)
-
-    def __str__(self):
-        return f"Image for {self.blog_post.title}"
-
