+++ apps/blog/models/author.py
@@ -1,0 +1,12 @@
+from django.db import models
+
+
+
+class Author(models.Model):
+    name = models.CharField(max_length=100)
+    bio = models.TextField()
+
+    def __str__(self):
+        return self.name
+
+
