+++ apps/blog/models/blog.py
@@ -1,0 +1,19 @@
+from django.db import models
+from apps.blog.models.author import Author
+from apps.blog.models.tag import Tag
+from apps.blog.models.category import Category
+
+class Blog(models.Model):
+    title = models.CharField(max_length=200)
+    content = models.TextField()
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+    author = models.ForeignKey(Author, on_delete=models.CASCADE)
+    categories = models.ManyToManyField(Category)
+    tags = models.ManyToManyField(Tag)
+
+    def __str__(self):
+        return self.title
+
+
+
