+++ apps/blog/models/category.py
@@ -1,0 +1,9 @@
+from django.db import models
+
+class Category(models.Model):
+    name = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.name
+
+
