+++ apps/blog/models/tag.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+class Tag(models.Model):
+    name = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.name
+
