+++ apps/blog/serializers/author_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from apps.blog.models.author import Author
+
+
+class AuthorSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Author
+        fields = ('id', 'name', 'bio')
