+++ apps/blog/serializers/category_serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from apps.blog.models.category import Category
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = '__all__'
