+++ apps/blog/serializers/image_serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from apps.blog.models.image import Image
+
+class ImageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Image
+        fields = '__all__'
