--- apps/blog/urls.py
@@ -1,8 +1,0 @@
-from django.urls import path
-from . import views
-
-urlpatterns = [
-    path('posts/', views.PostList.as_view(), name='post-list'),
-    path('posts/<int:pk>/', views.PostDetail.as_view(), name='post-detail'),
-    path('posts/create/', views.BlogListCreateView.as_view(), name='post-create'),
-]
