+++ apps/blog/views/index.py
@@ -1,0 +1,5 @@
+from django.http import HttpResponse
+
+
+def index(request):
+    return HttpResponse("Hello, world! This is the django cameroon page.")
