+++ apps/blog/views/post.py
@@ -1,14 +1,10 @@
 from rest_framework import generics, permissions
-from .models import Blog, Tag, Category, Author, Image
-from .serializers import BlogSerializer, BlogCreateUpdateSerializer, TagSerializer, CategorySerializer, AuthorSerializer, ImageSerializer
-
-from .serializers import BlogSerializer, BlogCreateUpdateSerializer, TagSerializer, CategorySerializer, AuthorSerializer, ImageSerializer
-
-
-from .tasks import handle_image_upload
-from django.http import HttpResponse
+from apps.blog.models.blog import Blog
+from apps.blog.serializers.blog_serializer import  BlogSerializer, BlogCreateUpdateSerializer
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
+
+
 class PostList(generics.ListCreateAPIView):
     queryset = Blog.objects.all()
 
@@ -66,43 +62,3 @@ class PostDetail(generics.RetrieveUpdateDestroyAPIView):
     )
     def delete(self, request, *args, **kwargs):
         return super().delete(request, *args, **kwargs)
-
-class TagListView(generics.ListCreateAPIView):
-    queryset = Tag.objects.all()
-    serializer_class = TagSerializer
-
-class CategoryListView(generics.ListCreateAPIView):
-    queryset = Category.objects.all()
-    serializer_class = CategorySerializer
-
-class AuthorListView(generics.ListCreateAPIView):
-    queryset = Author.objects.all()
-    serializer_class = AuthorSerializer
-
-class ImageListView(generics.ListCreateAPIView):
-    queryset = Image.objects.all()
-    serializer_class = ImageSerializer
-
-class ImageCreateView(generics.CreateAPIView):
-    queryset = Image.objects.all()
-    serializer_class = ImageSerializer
-
-    def perform_create(self, serializer):
-        instance = serializer.save()
-        handle_image_upload.delay(instance.id)
-
-class BlogListCreateView(generics.ListCreateAPIView):
-    queryset = Blog.objects.all()
-    serializer_class = BlogCreateUpdateSerializer
-
-    def get_permissions(self):
-        if self.request.method == 'POST':
-            self.permission_classes = [permissions.IsAuthenticated]
-        return super().get_permissions()
-
-class BlogDetail(generics.RetrieveUpdateDestroyAPIView):
-    queryset = Blog.objects.all()
-    serializer_class = BlogSerializer
-
-def index(request):
-    return HttpResponse("Hello, world! This is the django cameroon page.")
