+++ apps/events/test/__init__.py
@@ -1,2 +1,1 @@
-ÿþ 
-
+ÿþ
